Fix bug calculating token amounts
With ETH, we have to send the total amount of ETH and the contract
divides it up among the recipients.

With Tokens, we send the amount for each recipient, and the contract
sends that amount to each recipient.

Due to a bug, we were sending the total amount intended for all recipients,
then the contract was sending that amount to each recipient, so we were sending
N times more than we should have (where N is the number of recipients).